+++ RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 * Terminology System: Properly populate property definitions in expansions
+* Terminology System: Support for multiple servers
 * Validator: Major rework of tx and validation test cases associated with reworking terminology / validator interface and ongoing testing + reconciliation with Ontoserver
 major upgrade to validation - use terminology server to perform more logic, and standardise interface based on agreements with Vocab & Ontoserver
 * Validator: Improve slicing error message